serial: implement own protocol rather than using asyncio Stream
The stream made it impossible to do good echo detection as it couldn't
do .readatleast(n, timeout). With the serial object being its own
protocol, it can now do echo detection that was tested with both modes
(echo on and off) of a standalone RS485 USB adapter as well as the
FAM14.